third_party/blink/web_tests/traversal/acid3-test-2-expected.txt

Test of behavior NodeIterator when nodes are removed, from Acid3.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS count is 1
PASS nodea is nodeb
PASS count is 2
PASS nodea is nodeb
PASS count is 3
PASS nodea is nodeb
PASS count is 4
PASS nodea is nodeb
PASS count is 5
PASS nodea is nodeb
PASS count is 6
PASS nodea is nodeb
PASS count is 7
PASS nodea is nodeb
PASS count is 8
PASS nodea is nodeb
PASS count is 9
PASS nodea is nodeb
PASS count is 10
PASS nodea is nodeb
PASS count is 11
PASS nodea is nodeb
PASS count is 12
PASS nodea is nodeb
PASS count is 13
PASS nodea is nodeb
PASS count is 14
PASS nodea is nodeb
PASS count is 15
PASS nodea is nodeb
PASS count is 16
PASS nodea is nodeb
PASS count is 17
PASS nodea is nodeb
PASS successfullyParsed is true

TEST COMPLETE
